// src/services/api/generated/models/capability.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface Capability {
    /**
     * Unique identifier
     */
    'id': string;
    /**
     * Capability name
     */
    'name': string;
    /**
     * Detailed description of this capability
     */
    'description'?: string;
    /**
     * Whether this capability can be used
     */
    'is_active'?: boolean;
    'created_at': string;
    'updated_at': string;
}