// src/services/api/generated/models/name-enum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Test Management API
 * API for managing boards, tests, and execution.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * * `CMD` - Command Execution * `CAMERA` - Camera Testing * `DISPLAY` - Display Output * `SOUND` - Sound/Audio * `FILE` - File System * `SENSOR` - Sensor Reading * `GPIO` - GPIO Control * `ADC` - Analog-to-Digital * `NETWORK` - Network Connectivity
 */

export const NameEnum = {
    CMD: 'CMD',
    CAMERA: 'CAMERA',
    DISPLAY: 'DISPLAY',
    SOUND: 'SOUND',
    FILE: 'FILE',
    SENSOR: 'SENSOR',
    GPIO: 'GPIO',
    ADC: 'ADC',
    NETWORK: 'NETWORK'
} as const;

export type NameEnum = typeof NameEnum[keyof typeof NameEnum];